pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "me-gpt"
version = "0.1.5"
authors = [
    { name="Daniel McDonald", email="daniel@danielmcdonald.dev" },
]
description = "A python package to clone yourself with the help of LLMs"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai==0.27.8",
    "python-dotenv==1.0.0",
    "numpy",
]

[project.urls]
"Homepage" = "https://github.com/djm93dev/me_gpt"
"Bug Tracker" = "https://github.com/djm93dev/me_gpt/issues"